Backend/requirements.txt
fastapi==0.128.1
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
email-validator>=2.0.0
